Moved home COAR functionality to separate component
