+++ src/app/home-page/home-page.component.html
@@ -1,3 +1,4 @@
+<ds-home-coar></ds-home-coar>
 <ds-themed-home-news></ds-themed-home-news>
 <div [ngClass]="showDiscoverFilters ? 'container-fluid' : 'container'">
   <ds-page-with-sidebar [sidebarContent]="sidebar" [sideBarWidth]="showDiscoverFilters ? 3 : 0" [class]="showDiscoverFilters ? 'row mx-3' : ''">
